implement auto-submits
