+++ packages/eas-cli/src/commands/submit.ts
@@ -1,12 +1,11 @@
 import { getConfig } from '@expo/config';
-import { Platform } from '@expo/eas-build-job';
 import { EasJsonReader } from '@expo/eas-json';
 import { flags } from '@oclif/command';
 import { error } from '@oclif/errors';
 import chalk from 'chalk';
 
 import EasCommand from '../commandUtils/EasCommand';
-import { AppPlatform, SubmissionFragment, SubmissionStatus } from '../graphql/generated';
+import { SubmissionFragment } from '../graphql/generated';
 import { toAppPlatform } from '../graphql/types/AppPlatform';
 import Log, { learnMore } from '../log';
 import {
@@ -18,16 +17,9 @@ import {
 } from '../platform';
 import { isEasEnabledForProjectAsync, warnEasUnavailable } from '../project/isEasEnabledForProject';
 import { findProjectRootAsync, getProjectIdAsync } from '../project/projectUtils';
-import AndroidSubmitCommand from '../submissions/android/AndroidSubmitCommand';
-import {
-  SubmissionContext,
-  SubmitArchiveFlags,
-  createSubmissionContext,
-} from '../submissions/context';
-import IosSubmitCommand from '../submissions/ios/IosSubmitCommand';
-import { displayLogsAsync } from '../submissions/utils/logs';
+import { SubmitArchiveFlags, createSubmissionContext } from '../submissions/context';
+import { submitAsync, waitToCompleteAsync } from '../submissions/submit';
 import { printSubmissionDetailsUrls } from '../submissions/utils/urls';
-import { waitForSubmissionsEndAsync } from '../submissions/utils/wait';
 
 interface RawFlags {
   platform?: string;
@@ -135,7 +127,7 @@ See how to configure submits with eas.json: ${learnMore('https://docs.expo.dev/s
         );
       }
 
-      const submission = await this.submitAsync(ctx);
+      const submission = await submitAsync(ctx);
       submissions.push(submission);
     }
 
@@ -143,7 +135,7 @@ See how to configure submits with eas.json: ${learnMore('https://docs.expo.dev/s
     printSubmissionDetailsUrls(submissions);
 
     if (flags.wait) {
-      this.waitToCompleteAsync(submissions, flags);
+      await waitToCompleteAsync(submissions, { verbose: flags.verbose });
     }
   }
 
@@ -179,71 +171,5 @@ See how to configure submits with eas.json: ${learnMore('https://docs.expo.dev/s
       profile,
       nonInteractive,
     };
-  }
-
-  private async submitAsync<T extends Platform>(
-    ctx: SubmissionContext<T>
-  ): Promise<SubmissionFragment> {
-    const command =
-      ctx.platform === Platform.ANDROID
-        ? new AndroidSubmitCommand(ctx as SubmissionContext<Platform.ANDROID>)
-        : new IosSubmitCommand(ctx as SubmissionContext<Platform.IOS>);
-    return command.runAsync();
-  }
-
-  private async waitToCompleteAsync(
-    submissions: SubmissionFragment[],
-    flags: Flags
-  ): Promise<void> {
-    Log.newLine();
-    const completedSubmissions = await waitForSubmissionsEndAsync(submissions);
-    if (completedSubmissions.length > 1) {
-      Log.newLine();
-    }
-    for (const submission of completedSubmissions) {
-      if (completedSubmissions.length > 1) {
-        Log.log(
-          `${appPlatformEmojis[submission.platform]} ${chalk.bold(
-            `${appPlatformDisplayNames[submission.platform]} submission`
-          )}`
-        );
-      }
-      this.printInstructionsForIosSubmission(submission);
-      await displayLogsAsync(submission, { verbose: flags.verbose });
-      if (completedSubmissions.length > 1) {
-        Log.newLine();
-      }
-    }
-    this.exitWithNonZeroCodeIfSomeSubmissionsDidntFinish(completedSubmissions);
-  }
-
-  private printInstructionsForIosSubmission(submission: SubmissionFragment): void {
-    if (
-      submission.platform === AppPlatform.Ios &&
-      submission.status === SubmissionStatus.Finished
-    ) {
-      const logMsg = [
-        chalk.bold('Your binary has been successfully uploaded to App Store Connect!'),
-        '- It is now being processed by Apple - you will receive an e-mail when the processing finishes.',
-        '- It usually takes about 5-10 minutes depending on how busy Apple servers are.',
-        // ascAppIdentifier should be always available for ios submissions but check it anyway
-        submission.iosConfig?.ascAppIdentifier &&
-          `- When it’s done, you can see your build here: ${learnMore(
-            `https://appstoreconnect.apple.com/apps/${submission.iosConfig?.ascAppIdentifier}/appstore/ios`,
-            { learnMoreMessage: '' }
-          )}`,
-      ].join('\n');
-      Log.addNewLineIfNone();
-      Log.log(logMsg);
-    }
-  }
-
-  private exitWithNonZeroCodeIfSomeSubmissionsDidntFinish(submissions: SubmissionFragment[]): void {
-    const nonFinishedSubmissions = submissions.filter(
-      ({ status }) => status !== SubmissionStatus.Finished
-    );
-    if (nonFinishedSubmissions.length > 0) {
-      process.exit(1);
-    }
   }
 }
