+++ packages/eas-cli/src/submissions/android/__tests__/AndroidSubmitCommand-test.ts
@@ -156,7 +156,6 @@ describe(AndroidSubmitCommand, () => {
       expect(SubmissionMutation.createAndroidSubmissionAsync).toHaveBeenCalledWith({
         appId: projectId,
         config: {
-          archiveUrl: 'http://expo.dev/fake.apk',
           applicationIdentifier: testProject.appJSON.expo.android?.package,
           googleServiceAccountKeyJson: fakeFiles['/google-service-account.json'],
           releaseStatus: SubmissionAndroidReleaseStatus.Draft,
