+++ packages/eas-cli/src/submissions/utils/builds.ts
@@ -1,10 +1,5 @@
-import { AppPlatform, BuildArtifacts, BuildFragment, BuildStatus } from '../../graphql/generated';
+import { AppPlatform, BuildFragment, BuildStatus } from '../../graphql/generated';
 import { BuildQuery } from '../../graphql/queries/BuildQuery';
-
-// `BuildFragment` with non-null `artifacts.buildUrl`
-type BuildFragmentWithArtifact = BuildFragment & {
-  artifacts: BuildArtifacts & { buildUrl: string };
-};
 
 /**
  * Gets build by ID and ensures that `artifacts.buildUrl` exists
@@ -12,15 +7,11 @@ type BuildFragmentWithArtifact = BuildFragment & {
 export async function getBuildByIdForSubmissionAsync(
   platform: AppPlatform,
   buildId: string
-): Promise<BuildFragmentWithArtifact> {
+): Promise<BuildFragment> {
   const build = await BuildQuery.byIdAsync(buildId);
 
   if (build.platform !== platform) {
     throw new Error("Build platform doesn't match!");
-  }
-
-  if (!buildFragmentHasArtifact(build)) {
-    throw new Error('Build has no artifacts or build URL is not defined.');
   }
 
   return build;
@@ -29,28 +20,13 @@ export async function getBuildByIdForSubmissionAsync(
 export async function getLatestBuildForSubmissionAsync(
   platform: AppPlatform,
   appId: string
-): Promise<BuildFragmentWithArtifact | null> {
-  const builds = await BuildQuery.allForAppAsync(appId, {
+): Promise<BuildFragment | null> {
+  const [build] = await BuildQuery.allForAppAsync(appId, {
     limit: 1,
     filter: {
       platform,
       status: BuildStatus.Finished,
     },
   });
-
-  if (builds.length < 1) {
-    return null;
-  }
-
-  const build = builds[0];
-  if (!buildFragmentHasArtifact(build)) {
-    return null;
-  }
-
   return build;
-}
-
-// Utility function needed to make TypeScript happy
-function buildFragmentHasArtifact(build: BuildFragment): build is BuildFragmentWithArtifact {
-  return build.artifacts?.buildUrl != null;
 }
